+++ src/apps/ENGINE/src/compiler.cpp
@@ -309,8 +309,8 @@ void COMPILER::DTrace(const char *data_PTR, ...)
 
 // append one block of code to another
 // return new pointer
-/*char * COMPILER::AppendProgram(char * pBase_program, long Base_program_size, char * pAppend_program, long
-Append_program_size, long& new_program_size)
+/*char * COMPILER::AppendProgram(char * pBase_program, int32_t Base_program_size, char * pAppend_program, int32_t
+Append_program_size, int32_t& new_program_size)
 {
     char * pTemp;
     char * pBase;
@@ -595,7 +595,7 @@ void COMPILER::FindErrorSource()
     } while (Token_type != END_OF_PROGRAMM);
 }
 
-void COMPILER::SetEventHandler(const char *event_name, const char *func_name, long flag, bool bStatic)
+void COMPILER::SetEventHandler(const char *event_name, const char *func_name, int32_t flag, bool bStatic)
 {
     FuncInfo fi;
 
@@ -649,7 +649,7 @@ void COMPILER::DelEventHandler(const char *event_name, const char *func_name)
 
     EventTab.SetStatus(event_name, func_code, FSTATUS_DELETED);
 
-    for (long n = 0; n < static_cast<long>(EventMsg.GetClassesNum()); n++)
+    for (int32_t n = 0; n < static_cast<int32_t>(EventMsg.GetClassesNum()); n++)
     {
         S_EVENTMSG *pM = EventMsg.Read(n);
         if (!pM->pEventName)
@@ -1083,7 +1083,7 @@ void COMPILER::ProcessFrame(uint32_t DeltaTime)
 
     EventTab.ProcessFrame();
 
-    for (long ln = 0; ln < static_cast<long>(EventMsg.GetClassesNum()); ln++)
+    for (int32_t ln = 0; ln < static_cast<int32_t>(EventMsg.GetClassesNum()); ln++)
     {
         S_EVENTMSG *pMsg = EventMsg.Read(ln);
         if (pMsg->bInvalide)
@@ -1267,7 +1267,7 @@ bool COMPILER::Compile(SEGMENT_DESC &Segment, char *pInternalCode, uint32_t pInt
 
     //    bool bCDStop;
     bool bFunctionBlock;
-    long lvalue;
+    int32_t lvalue;
     float fvalue;
 
     Control_offset = 0;
@@ -1447,10 +1447,10 @@ bool COMPILER::Compile(SEGMENT_DESC &Segment, char *pInternalCode, uint32_t pInt
                 switch (Token.GetType())
                 {
                 case NUMBER:
-                    lvalue = static_cast<long>(atoll(Token.GetData()));
+                    lvalue = static_cast<int32_t>(atoll(Token.GetData()));
                     if (bMinus)
                         lvalue = -lvalue;
-                    memcpy(&di.data4b, &lvalue, sizeof(long));
+                    memcpy(&di.data4b, &lvalue, sizeof(int32_t));
                     break;
                 case FLOAT_NUMBER:
                     fvalue = static_cast<float>(atof(Token.GetData()));
@@ -1602,7 +1602,7 @@ bool COMPILER::Compile(SEGMENT_DESC &Segment, char *pInternalCode, uint32_t pInt
                                 }
                             }
                             else
-                                lvalue = static_cast<long>(atoll(Token.GetData()));
+                                lvalue = static_cast<int32_t>(atoll(Token.GetData()));
                             vi.elements = lvalue;
                             Token.Get(); // SQUARE_CLOSE_BRACKET
                             Token.Get();
@@ -1835,7 +1835,7 @@ bool COMPILER::Compile(SEGMENT_DESC &Segment, char *pInternalCode, uint32_t pInt
                             }
                         }
                         else
-                            lvalue = static_cast<long>(atoll(Token.GetData()));
+                            lvalue = static_cast<int32_t>(atoll(Token.GetData()));
                         lvi.elements = lvalue;
                         Token.Get(); // SQUARE_CLOSE_BRACKET
                         Token.Get();
@@ -2072,7 +2072,7 @@ bool COMPILER::CompileBlock(SEGMENT_DESC &Segment, bool &bFunctionBlock, uint32_
     BreakUpdateTable.SetStringDataSize(sizeof(uint32_t));
     bool bExtern;
     bool bImport;
-    long lvalue;
+    int32_t lvalue;
     uint32_t def_code;
     DEFINFO di;
     uint32_t dwRCode;
@@ -3439,11 +3439,11 @@ bool COMPILER::CompileBlock(SEGMENT_DESC &Segment, bool &bFunctionBlock, uint32_
 
         case TRUE_CONST:
             lvalue = 1;
-            CompileToken(Segment, NUMBER, 1, (char *)&lvalue, sizeof(long));
+            CompileToken(Segment, NUMBER, 1, (char *)&lvalue, sizeof(int32_t));
             break;
         case FALSE_CONST:
             lvalue = 0;
-            CompileToken(Segment, NUMBER, 1, (char *)&lvalue, sizeof(long));
+            CompileToken(Segment, NUMBER, 1, (char *)&lvalue, sizeof(int32_t));
             break;
 
             // write value
@@ -3576,7 +3576,7 @@ bool COMPILER::BC_CallFunction(uint32_t func_code, uint32_t &ip, DATA *&pVResult
         SetError("missing args_num token");
         return false;
     }
-    arguments = *((long *)&pRunCodeBase[TLR_DataOffset]);
+    arguments = *((int32_t *)&pRunCodeBase[TLR_DataOffset]);
 
     check_sp = SStack.GetDataNum() - arguments;
     /*
@@ -3755,7 +3755,7 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
     DATA *pVV;
     uint32_t var_code;
     const char *pAccess_string;
-    //    long array_index;
+    //    int32_t array_index;
     //    DATA * pRef;
     DATA *pVar;
     entid_t eid;
@@ -3945,13 +3945,13 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
         case LOCAL_VARIABLE:
             pLeftOperandAClass = nullptr;            // reset attribute
             nLeftOperandIndex = INVALID_ARRAY_INDEX; // reset index
-            nLeftOperandCode = *((long *)&pRunCodeBase[TLR_DataOffset]);
+            nLeftOperandCode = *((int32_t *)&pRunCodeBase[TLR_DataOffset]);
             bLeftOperandType = LOCAL_VARIABLE;
             break;
         case VARIABLE:
             pLeftOperandAClass = nullptr;            // reset attribute
             nLeftOperandIndex = INVALID_ARRAY_INDEX; // reset index
-            nLeftOperandCode = *((long *)&pRunCodeBase[TLR_DataOffset]);
+            nLeftOperandCode = *((int32_t *)&pRunCodeBase[TLR_DataOffset]);
             bLeftOperandType = VARIABLE;
             break;
         case ACCESS_WORD_CODE:
@@ -3993,7 +3993,7 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
                     return false;
                 }
                 pLeftOperandAClass =
-                    pLeftOperandAClass->VerifyAttributeClassByCode(*((long *)&pRunCodeBase[TLR_DataOffset]));
+                    pLeftOperandAClass->VerifyAttributeClassByCode(*((int32_t *)&pRunCodeBase[TLR_DataOffset]));
                 break;
             }
             if (pLeftOperandAClass == nullptr)
@@ -4026,7 +4026,7 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
                 return false;
             }
             pLeftOperandAClass =
-                pLeftOperandAClass->VerifyAttributeClassByCode(*((long *)&pRunCodeBase[TLR_DataOffset]));
+                pLeftOperandAClass->VerifyAttributeClassByCode(*((int32_t *)&pRunCodeBase[TLR_DataOffset]));
             break;
             break;
         case ACCESS_WORD:
@@ -4140,7 +4140,7 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
                 }
 
                 vtype = BC_TokenGet();
-                var_code = *((long *)&pRunCodeBase[TLR_DataOffset]);
+                var_code = *((int32_t *)&pRunCodeBase[TLR_DataOffset]);
                 if (!(vtype == VARIABLE || vtype == LOCAL_VARIABLE))
                 {
                     SetError("invalid access var");
@@ -4203,7 +4203,7 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
             }
 
             vtype = BC_TokenGet();
-            var_code = *((long *)&pRunCodeBase[TLR_DataOffset]);
+            var_code = *((int32_t *)&pRunCodeBase[TLR_DataOffset]);
             if (!(vtype == VARIABLE || vtype == LOCAL_VARIABLE))
             {
                 SetError("invalid access var");
@@ -4303,9 +4303,9 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
 
             /*if(!bTraceMode) break;
 
-            long index;
-            long file_line_offset;
-            long file_name_size;
+            int32_t index;
+            int32_t file_line_offset;
+            int32_t file_name_size;
             index = ip;
             memcpy(&file_line_offset,&pCodeBase[index],sizeof(uint32_t)); index += sizeof(uint32_t);
             memcpy(&file_name_size,&pCodeBase[index],sizeof(uint32_t)); index += sizeof(uint32_t);
@@ -4325,7 +4325,7 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
 
         case CALL:                                               // undetermined function call
             vtype = BC_TokenGet();                               // read variable
-            var_code = *((long *)&pRunCodeBase[TLR_DataOffset]); // var code
+            var_code = *((int32_t *)&pRunCodeBase[TLR_DataOffset]); // var code
             if (vtype == VARIABLE)
             {
                 real_var = VarTab.GetVar(var_code);
@@ -4819,7 +4819,7 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
               {
                 if(TokenType() == ACCESS_WORD_CODE)
                 {
-                  pRoot = pRoot->VerifyAttributeClassByCode(*((long *)&pRunCodeBase[TLR_DataOffset]));
+                  pRoot = pRoot->VerifyAttributeClassByCode(*((int32_t *)&pRunCodeBase[TLR_DataOffset]));
                   // pRoot = pRoot->CreateSubAClass(pRoot,(char *)&pRunCodeBase[TLR_DataOffset]);
                 } else
                 if(TokenType() == ACCESS_WORD)
@@ -4829,7 +4829,7 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
                 if(TokenType() == ACCESS_VAR)
                 {
                   vtype = BC_TokenGet();
-                  var_code = *((long *)&pRunCodeBase[TLR_DataOffset]);
+                  var_code = *((int32_t *)&pRunCodeBase[TLR_DataOffset]);
                   if(!(vtype == VARIABLE || vtype == LOCAL_VARIABLE)){SetError("invalid access var");return false;}
                   if(vtype == VARIABLE)
                   {
@@ -5626,14 +5626,14 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
             {
             case ACCESS_WORD_CODE:
                 if (sttV == VERIFY_AP)
-                    rAP = rAP->VerifyAttributeClassByCode(*((long *)&pRunCodeBase[TLR_DataOffset]));
+                    rAP = rAP->VerifyAttributeClassByCode(*((int32_t *)&pRunCodeBase[TLR_DataOffset]));
                 else
-                    rAP = rAP->GetAttributeClassByCode(*((long *)&pRunCodeBase[TLR_DataOffset]));
+                    rAP = rAP->GetAttributeClassByCode(*((int32_t *)&pRunCodeBase[TLR_DataOffset]));
                 if (!rAP)
-                    SetError("missed attribute: %s", SCodec.Convert(*((long *)&pRunCodeBase[TLR_DataOffset])));
+                    SetError("missed attribute: %s", SCodec.Convert(*((int32_t *)&pRunCodeBase[TLR_DataOffset])));
                 break;
             case VARIABLE:
-                real_var = VarTab.GetVar(*((long *)&pRunCodeBase[TLR_DataOffset]));
+                real_var = VarTab.GetVar(*((int32_t *)&pRunCodeBase[TLR_DataOffset]));
                 if (real_var == nullptr)
                 {
                     SetError("Global variable not found");
@@ -5661,7 +5661,7 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
                     SetError("missed attribute: %s", pChar);
                 break;
             case LOCAL_VARIABLE:
-                pV = SStack.Read(pRun_fi->stack_offset, *((long *)&pRunCodeBase[TLR_DataOffset]));
+                pV = SStack.Read(pRun_fi->stack_offset, *((int32_t *)&pRunCodeBase[TLR_DataOffset]));
                 if (pV == nullptr)
                 {
                     SetError("Local variable not found");
@@ -5713,9 +5713,9 @@ bool COMPILER::BC_Execute(uint32_t function_code, DATA *&pVReturnResult, const c
 
 void COMPILER::CompileNumber(SEGMENT_DESC &Segment)
 {
-    long lvalue;
-    lvalue = static_cast<long>(atoll(Token.GetData()));
-    CompileToken(Segment, NUMBER, 1, (char *)&lvalue, sizeof(long));
+    int32_t lvalue;
+    lvalue = static_cast<int32_t>(atoll(Token.GetData()));
+    CompileToken(Segment, NUMBER, 1, (char *)&lvalue, sizeof(int32_t));
 }
 
 void COMPILER::CompileFloatNumber(SEGMENT_DESC &Segment)
@@ -6049,7 +6049,7 @@ ATTRIBUTES *COMPILER::TraceARoot(ATTRIBUTES *pA, const char *&pAccess)
     if (pA->GetThisNameCode() == 0)
         return nullptr; // fix crash at NewGame start
 
-    const long slen = strlen(pA->GetThisName()) + 1;
+    const int32_t slen = strlen(pA->GetThisName()) + 1;
 
     char *pAS = new char[slen];
 
@@ -6239,7 +6239,7 @@ bool COMPILER::ReadVariable(char *name, /* DWORD code,*/ bool bDim, uint32_t a_i
             if (bSkipVariable)
             {
                 if (eType == S_TOKEN_TYPE::VAR_INTEGER)
-                    ReadData(nullptr, sizeof(long));
+                    ReadData(nullptr, sizeof(int32_t));
                 else if (eType == S_TOKEN_TYPE::VAR_FLOAT)
                     ReadData(nullptr, sizeof(float));
                 else if (eType == S_TOKEN_TYPE::VAR_STRING)
@@ -6760,7 +6760,7 @@ void COMPILER::AddPostEvent(S_EVENTMSG *pEM)
     EventMsg.Add(pEM);
 }
 
-bool COMPILER::SetSaveData(const char *file_name, void *save_data, long data_size)
+bool COMPILER::SetSaveData(const char *file_name, void *save_data, int32_t data_size)
 {
     EXTDATA_HEADER exdh;
 
@@ -6795,7 +6795,7 @@ bool COMPILER::SetSaveData(const char *file_name, void *save_data, long data_siz
     return true;
 }
 
-/*bool COMPILER::SetSaveData(char * file_name, void * save_data, long data_size)
+/*bool COMPILER::SetSaveData(char * file_name, void * save_data, int32_t data_size)
                            //const char * file_name, const char * save_data)
 {
     EXTDATA_HEADER exdh;
@@ -6881,7 +6881,7 @@ bool COMPILER::SetSaveData(const char *file_name, void *save_data, long data_siz
 
 }*/
 
-void *COMPILER::GetSaveData(const char *file_name, long &data_size)
+void *COMPILER::GetSaveData(const char *file_name, int32_t &data_size)
 {
     auto fileS = fio->_CreateFile(file_name, std::ios::binary | std::ios::in);
     if (!fileS.is_open())
@@ -6932,7 +6932,7 @@ void *COMPILER::GetSaveData(const char *file_name, long &data_size)
     return pBuffer;
 }
 
-/*void * COMPILER::GetSaveData(char * file_name, long & data_size)
+/*void * COMPILER::GetSaveData(char * file_name, int32_t & data_size)
 {
     DWORD n;
